Move Zookeeper code into lib
