+++ src/lib/zookeeper/components/MlEphantConversationPaneHooks.ts
@@ -2,8 +2,8 @@ import { type MlToolResult } from '@kittycad/lib'
 import type { FileEntry } from '@src/lib/project'
 import type { SettingsType } from '@src/lib/settings/initialSettings'
 import { reportRejection } from '@src/lib/trap'
-import type { ZookeeperConversationStore } from '@src/lib/zookeeperConversationStore'
-import { type MlEphantManagerActor } from '@src/machines/mlEphantManagerMachine'
+import type { ZookeeperConversationStore } from '@src/lib/zookeeper/zookeeperConversationStore'
+import { type MlEphantManagerActor } from '@src/lib/zookeeper/mlEphantManagerMachine'
 import { type RequestedKCLFileDelete } from '@src/machines/systemIO/utils'
 import type { ConnectionManager } from '@src/network/connectionManager'
 import { useEffect } from 'react'
