+++ src/lib/zookeeper/registry/index.spec.ts
@@ -13,9 +13,12 @@ import { layoutAreaLibraryValueSpec } from '@src/registry/contracts/layout'
 import { statusBarLocalItemsValueSpec } from '@src/registry/contracts/statusBar'
 import { describe, expect, it, vi } from 'vitest'
 
-vi.mock('@src/components/layout/areas/MlEphantConversationPaneWrapper', () => ({
-  MlEphantConversationPaneWrapper: () => null,
-}))
+vi.mock(
+  '@src/lib/zookeeper/components/MlEphantConversationPaneWrapper',
+  () => ({
+    MlEphantConversationPaneWrapper: () => null,
+  })
+)
 
 vi.mock('@src/components/ZookeeperCreditsMenu', () => ({
   ZookeeperCreditsMenu: () => null,
