+++ src/registry/registry.ts
@@ -6,9 +6,16 @@ type RegistryItemModule = {
 }
 
 const bundledRegistryItemModules: Record<string, RegistryItemModule> =
-  import.meta.glob(['./extensions/*/index.ts', './plugins/*/index.ts'], {
-    eager: true,
-  })
+  import.meta.glob(
+    [
+      './extensions/*/index.ts',
+      './plugins/*/index.ts',
+      '../lib/*/registry/index.ts',
+    ],
+    {
+      eager: true,
+    }
+  )
 
 export const appRegistryServicesSlot = new Slot()
 export const appRegistryOverridesSlot = new Slot()
@@ -16,8 +23,8 @@ export const appRegistryOverridesSlot = new Slot()
 const bundledRegistryItemKindOrder = (path: string) =>
   path.includes('/plugins/') ? 1 : 0
 
-// Core app registry items discovered from src/registry/extensions/*/index.ts
-// and src/registry/plugins/*/index.ts.
+// Core app registry items discovered from src/registry/extensions/*/index.ts,
+// src/registry/plugins/*/index.ts, and src/lib/*/registry/index.ts.
 //
 // Extensions are always-on app infrastructure. Plugins are user-visible,
 // runtime-toggleable bundles that may depend on extension contracts.
